backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
filelock==3.24.3
fsspec==2026.2.0
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.62.3
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==1.26.4
packaging==26.0
pgvector==0.3.2
pillow==10.4.0
proto-plus==1.27.1
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.5
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.6.0
